fix: colab上でtestフォルダが参照できない問題への対応
