+++ tests/test_not_broken_various_inputs_type2.py
@@ -4,7 +4,7 @@ import copy
 
 from jjjexperiment.main import calc
 
-from test_utils.util import  \
+from test_utils.expects import  \
     expected_inputs, expected_result_type2
 
 class Test既存計算維持_入力値切替_方式2:
